[Oztechan/CCC#4256] add support for android 15
