+++ android/ui/mobile/src/main/kotlin/com/oztechan/ccc/android/ui/mobile/content/main/MainActivity.kt
@@ -8,11 +8,9 @@ import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import android.view.View
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.net.toUri
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.core.view.ViewCompat
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import co.touchlab.kermit.Logger
@@ -20,6 +18,7 @@ import com.github.submob.basemob.activity.BaseActivity
 import com.oztechan.ccc.android.core.ad.AdManager
 import com.oztechan.ccc.android.ui.mobile.BuildConfig
 import com.oztechan.ccc.android.ui.mobile.R
+import com.oztechan.ccc.android.ui.mobile.util.applyWindowInsets
 import com.oztechan.ccc.android.ui.mobile.util.getThemeMode
 import com.oztechan.ccc.android.ui.mobile.util.isDeviceRooted
 import com.oztechan.ccc.android.ui.mobile.util.requestAppReview
@@ -33,9 +32,6 @@ import com.oztechan.ccc.client.viewmodel.main.MainViewModel
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import org.koin.android.ext.android.inject
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.updatePadding
-
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainActivity : BaseActivity() {
@@ -51,19 +47,22 @@ class MainActivity : BaseActivity() {
         } // else is in on create since viewModel needs to be injected
     }
 
+    @Suppress("DEPRECATION")
     override fun onCreate(savedInstanceState: Bundle?) {
-        enableEdgeToEdge()
+        //enableEdgeToEdge()
         installSplashScreen()
+
+        window.navigationBarColor = getColor(R.color.background_strong)
+        window.statusBarColor = getColor(R.color.background_strong)
+        actionBar?.hide()
 
         super.onCreate(savedInstanceState)
 
         Logger.i { "MainActivity onCreate" }
 
-        setTitle("")
         setContentView(R.layout.activity_main)
 
-        val navHostFragment = findViewById<View>(R.id.content)
-        navHostFragment.applyWindowInsets()
+        findViewById<View>(R.id.content).applyWindowInsets()
 
         adManager.initAds(this)
         analyticsManager.setUserProperty(UserProperty.IsRooted(isDeviceRooted(this)))
@@ -136,26 +135,5 @@ class MainActivity : BaseActivity() {
 
     override fun attachBaseContext(base: Context) {
         super.attachBaseContext(updateBaseContextLocale(base))
-    }
-}
-
-fun View.applyWindowInsets(
-    applyTop: Boolean = true,
-    applyBottom: Boolean = true,
-    applyLeft: Boolean = true,
-    applyRight: Boolean = true
-) {
-    ViewCompat.setOnApplyWindowInsetsListener(this) { view, windowInsets ->
-        val systemInsets = windowInsets.getInsets(WindowInsetsCompat.Type.systemBars())
-        val cutoutInsets = windowInsets.getInsets(WindowInsetsCompat.Type.displayCutout())
-
-        view.updatePadding(
-            left = if (applyLeft) maxOf(systemInsets.left, cutoutInsets.left) else 0,
-            top = if (applyTop) maxOf(systemInsets.top, cutoutInsets.top) else 0,
-            right = if (applyRight) maxOf(systemInsets.right, cutoutInsets.right) else 0,
-            bottom = if (applyBottom) maxOf(systemInsets.bottom, cutoutInsets.bottom) else 0
-        )
-
-        windowInsets
     }
 }
