+++ android/ui/mobile/src/main/res/values/styles.xml
@@ -18,7 +18,7 @@
         <item name="windowNoTitle">true</item>
         <item name="android:windowAnimationStyle">@style/AppTheme.WindowTransition</item>
         <item name="android:actionOverflowButtonStyle">@style/ActionButtonOverflow</item>
-        <item name="android:navigationBarColor">@android:color/transparent</item>
+        <item name="android:navigationBarColor">@color/background_strong</item>
     </style>
 
     <style name="AppTheme" parent="BaseAppTheme" />
